Thread in DataGenerator waits for a 3000 msek before it prints the content again
