+++ app/src/main/java/com/example/eliasvensson/busify/DataGenerator.java
@@ -29,7 +29,7 @@ public class DataGenerator {
      * and the String containing the date for which bus information is to be shown.
      */
     private FirebaseDatabase database;
-    private DatabaseReference ref;
+    protected DatabaseReference ref;
     private String chosenDate;
     private Activity mainActivity;
     private String busdata = "";
@@ -65,18 +65,6 @@ public class DataGenerator {
 
             @Override
             public void onDataChange(DataSnapshot dataSnapshot) {
-                busdata = dataSnapshot.getValue().toString();
-                // Gets a snapshot of the data in the database as a String
-                    Thread timerThread = new Thread() {
-                        public void run() {
-                            try {
-                                //Sets the duration of the splash screen
-                                sleep(3000);
-                            } catch (InterruptedException e) {
-                                e.printStackTrace();
-                            }
-                        }
-                    };
                 busdata = dataSnapshot.getValue().toString();
             }
