+++ app/src/main/java/com/example/eliasvensson/busify/SplashScreen.java
@@ -3,12 +3,6 @@ package com.example.eliasvensson.busify;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-import android.os.Handler;
-
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.MutableData;
-import com.google.firebase.database.Transaction;
 
 /**
  * @author Jonathan Fager
